合并最新代码

+++ src/main/java/io/mycat/net/ThreadLocalBufferPool.java
@@ -2,9 +2,9 @@ package io.mycat.net;
 
 
 public class ThreadLocalBufferPool extends ThreadLocal<BufferQueue> {
-	private final int size;
+	private final long size;
 
-	public ThreadLocalBufferPool(int size) {
+	public ThreadLocalBufferPool(long size) {
 		this.size = size;
 	}
